Moved settings logic to SettingsViewController and resolved conflicts
