+++ ISS Barometer/SettingsViewController.swift
@@ -15,6 +15,11 @@ class SettingsViewController: UITableViewController {
     @IBOutlet weak var sigFigValue: UILabel!
     @IBOutlet weak var unitPicker: UISegmentedControl!
     @IBOutlet weak var orientationPicker: UISegmentedControl!
+    @IBOutlet weak var slidingScalePicker: UISwitch!
+    @IBOutlet weak var runningWindowSlider: UISlider!
+    @IBOutlet weak var runningWindowValue: UILabel!
+    @IBOutlet weak var runningWindowOptions: UIView!
+    @IBOutlet weak var runningWindowTable: UITableViewCell!
     
     @IBAction func sliderMoved(_ sender: Any) {
         let roundedValue = lroundf(sigFigSlider.value)
@@ -36,6 +41,18 @@ class SettingsViewController: UITableViewController {
         UIDevice.current.setValue(selectedIdx + 1, forKey: "orientation")
     }
     
+    @IBAction func slidingScalePicked(_ sender: Any) {
+        settings?.slidingScale = slidingScalePicker.isOn
+        initWindowSlider()
+    }
+    
+    @IBAction func runningWindowSliderMoved(_ sender: Any) {
+        let roundedValue = lroundf(runningWindowSlider.value)
+        (sender as AnyObject).setValue(Float(roundedValue), animated: true)
+        runningWindowValue.text = String(roundedValue*25)
+        settings?.runningWindowSize = roundedValue*25
+    }
+    
     func initSlider() {
         sigFigValue.text = String(describing: settings!.sigFigs)
         sigFigSlider.setValue(Float(settings!.sigFigs), animated: true)
@@ -53,11 +70,29 @@ class SettingsViewController: UITableViewController {
         orientationPicker.selectedSegmentIndex = segmentIdx
     }
     
+    func initSlidingScalePicker() {
+        slidingScalePicker.setOn((settings?.slidingScale)!, animated: true)
+    }
+    
+    func initWindowSlider() {
+        if slidingScalePicker.isOn {
+            runningWindowOptions.isHidden = false
+            runningWindowTable.isHidden = false
+        } else {
+            runningWindowOptions.isHidden = true
+            runningWindowTable.isHidden = true
+        }
+        runningWindowValue.text = String(describing: settings!.runningWindowSize)
+        runningWindowSlider.setValue(Float(round(Double(settings!.runningWindowSize)/25.0)), animated: true)
+    }
+    
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
         initSlider()
         initUnitPicker()
         initOrientationPicker()
+        initSlidingScalePicker()
+        initWindowSlider()
     }
     
     override func viewDidLoad() {
